+++ bower.json
@@ -28,7 +28,7 @@
     "cal-heatmap": "~3.5.2",
     "moment": "~2.9.0",
     "numeral": "~1.5.3",
-    "heatmap.js.bower": "*",
+    "heatmap.js": "pa7/heatmap.js#develop",
     "datatables-bootstrap3-plugin": "~0.4.0",
     "dota2-minimap-hero-sprites": "bontscho/dota2-minimap-hero-sprites"
   }
