+++ package.json
@@ -16,7 +16,7 @@
     "getmissing": "node runner.js getmissing",
     "bower": "bower install",
     "webpack": "webpack -p",
-    "watch": "webpack watch",
+    "watch": "webpack --watch",
     "build": "npm install && npm run bower && npm run constants && npm run webpack && npm run maven",
     "maven": "mvn -q -f parser/pom.xml clean install -U",
     "yasp-parser": "sudo nf export -j Procfile.parser -o /etc/init/ && sudo restart foreman",
@@ -33,8 +33,12 @@
     "big-number": "0.3.1",
     "body-parser": "1.12.2",
     "bower": "^1.4.1",
+    "c3": "^0.4.10",
+    "cal-heatmap": "^3.5.2",
     "compression": "^1.4.3",
     "connect-redis": "^2.2.0",
+    "d3": "^3.5.5",
+    "datatables": "^1.10.6",
     "dota2": "yasp-dota/node-dota2",
     "dotenv": "~1.1.0",
     "express": "~4.12.3",
@@ -42,14 +46,17 @@
     "foreman": "~1.3.0",
     "http-auth": "~2.2.5",
     "jade": "~1.9.2",
+    "jquery": "^2.1.3",
     "kue": "~0.8.12",
-    "moment": "~2.10.2",
+    "moment": "^2.10.2",
     "monk": "1.0.1",
+    "numeral": "^1.5.3",
     "parse-redis-url": "0.0.2",
     "passport": "0.2.1",
     "passport-steam": "0.1.4",
     "paypal-rest-sdk": "~1.5",
     "poet": "~2.0.0",
+    "qtip2": "^2.2.0",
     "redis": "0.12.1",
     "request": "~2.55.0",
     "request-progress": "^0.3.1",
@@ -62,7 +69,9 @@
   },
   "devDependencies": {
     "coveralls": "2.11.2",
+    "css-loader": "^0.11.0",
     "database-cleaner": "~0.9.1",
+    "file-loader": "^0.8.1",
     "grunt": "^0.4.5",
     "grunt-cli": "^0.1.13",
     "grunt-contrib-cssmin": "^0.12.2",
@@ -74,7 +83,9 @@
     "mongodb": "~2.0.27",
     "nock": "~1.6.0",
     "nodemon": "^1.3.7",
+    "style-loader": "^0.11.0",
     "supertest": "^0.15.0",
+    "url-loader": "^0.5.5",
     "webpack": "^1.8.9"
   }
 }
