+++ public/js/histograms.js
@@ -1,34 +1,4 @@
-function createCalHeatmap(data) {
-    //create cal-heatmap from start_time
-    var cal = new CalHeatMap();
-    cal.init({
-        start: new Date(moment().subtract(1, 'year')),
-        itemSelector: "#cal-heatmap",
-        range: 13,
-        domain: "month",
-        subDomain: "day",
-        data: data,
-        verticalOrientation: true,
-        label: {
-            position: "left"
-        },
-        colLimit: 31,
-        tooltip: true,
-        legend: [1, 2, 3, 4],
-        highlight: new Date(),
-        itemName: ["match", "matches"],
-        subDomainTextFormat: function(date, value) {
-            return value;
-        },
-        cellSize: 13,
-        domainGutter: 5,
-        previousSelector: "#prev",
-        nextSelector: "#next",
-        legendHorizontalPosition: "right"
-    });
-}
-
-function createHistogram(counts, win_counts, label) {
+module.exports = function createHistogram(counts, win_counts, label) {
     //counts, the number of matches with a value in each key
     //win_counts, the number of wins with a value in each key
     //create a hash of the categories/buckets to wins/games data
