+++ public/js/playerMatches.js
@@ -1,4 +1,4 @@
-function playerMatches(teammates) {
+module.exports = function playerMatches(teammates) {
     //extend jquery to serialize form data to JSON
     $.fn.serializeObject = function() {
         var o = {};
