+++ public/js/ratingsChart.js
@@ -1,4 +1,4 @@
-function ratingsChart(ratings) {
+module.exports = function ratingsChart(ratings) {
     $("#solo").append(ratings[0] ? ratings[0].soloCompetitiveRank : "N/A");
     $("#party").append(ratings[0] ? ratings[0].competitiveRank : "N/A");
     var times = ratings.map(function(r) {
